Renaming panes to windows - fix https://github.com/facebookresearch/visdom/issues/162
Summary:
To clean up inconsistency in the docs, this standardizes all visdom calls to use "window" instead of "pane" to refer to the panels that contain graphs or other data. On the UI side these are still referred to as Panes, but for most users this isn't a discrepancy.

We may want to update this in the future to make things clearer for contributers

Also changes some string formatting to deal with errors that popped up when logging in the demos

Reviewed By: lvdmaaten

Differential Revision: D5986451

fbshipit-source-id: 236d56c3730e8fa692e8b5721866ba510aea1087