Window Missing Name Fix
Summary:
A recent change made it so that the server would return "None" (the string) as the name of any window that was not explicitly given a name, rather than generating a unique name like it was supposed to. This occurred because the arguments were recently forced to instantiate the window name to None (based on default args) and we didn't have a check to notice that it happened.

This PR fixes the problem and adds a check to our python demo
Closes https://github.com/facebookresearch/visdom/pull/182

Reviewed By: lvdmaaten

Differential Revision: D6172361

Pulled By: JackUrb

fbshipit-source-id: f84bdb24496949953effdb79195a0c6a3c4ec872