Legend Update, opts fix
Summary:
Turns out some of the opts we wanted to send were being removed by scrub_dict when they were false, which was not the behavior wanted for options that should in fact be false. Legends were also being handled inconsistently in general, so this standardizes using legends and toggling if they appear or not.

Fixes https://github.com/facebookresearch/visdom/issues/186

Tested using modifications to the demo
Closes https://github.com/facebookresearch/visdom/pull/187

Reviewed By: lvdmaaten

Differential Revision: D6197019

Pulled By: JackUrb

fbshipit-source-id: 49988aeecf52b4272f7b8dd3768dcf32d994e336