Stabilize pack
Summary:
Before this change, items being packed would often fly around the page with little to no control of what was going on. This fixes the sort bug that was causing this issue.

This change also ensures that server-side resizes are captured on the frontend when they happen by catching the change in a form that forces a React ComponentUpdate.

Fixes https://github.com/facebookresearch/visdom/issues/134.
Closes https://github.com/facebookresearch/visdom/pull/235

Differential Revision: D6642957

Pulled By: JackUrb

fbshipit-source-id: 04f1d9af87ac78bfeb7a1ecae3fc5224684e2b06