Adds functionality to delete traces from line and scatter plots
Summary:
This adds a new possible update type (remove) which allows users to remove a trace programatically.  Using `update='remove'` with a named trace and window will delete that trace from the window for both scatter and line plots.

Adds an example in the demo of using this feature

Fixes an old bug that was preventing new traces from receiving their specified names (and would all just be named '1') as it was likely causing issues for power-append users who would use the same window over and over for things.

Closes https://github.com/facebookresearch/visdom/issues/221 by implementing the feature.
Closes https://github.com/facebookresearch/visdom/pull/247

Differential Revision: D6702548

Pulled By: JackUrb

fbshipit-source-id: d8e2ab4be41f27b02deba8eb8932a81171bb3df1