Improves the filter to react to typing on the fly rather than using the button.
Summary:
This makes the filter bar active rather than passive. Typing will filter content immediately as you type. As long as you have saved your current view, deleting content from the filter bar will return your view to how it was.

The filter button was replaced with a button that clears the content of the filter.

This also fixes the bug that would occasionally cause plots to zoom out to as tiny as possible randomly and irreproducibly when using the old filter.

![betterfilter](https://user-images.githubusercontent.com/1276867/34849912-f0693f30-f6f1-11e7-90b6-2a39f83280e8.gif)

TODO add documentation about the new filter, attach the above gif to it
Closes https://github.com/facebookresearch/visdom/pull/248

Differential Revision: D6711549

Pulled By: JackUrb

fbshipit-source-id: 21f9acffbd257a18253fd78cbc28909e83bbeb70