Adds a way to save views to the server to load later
Summary:
Previously to this commit, views were lost on refresh. This allows for views to be synced with the server so that they can be used later for any number of reasons. Also moves the views feature out of *beta* now that it has at least the minimum planned functionality implemented.
Closes https://github.com/facebookresearch/visdom/pull/249

Differential Revision: D6737087

Pulled By: JackUrb

fbshipit-source-id: 9f3e1ce048d27a77dec1be073df4f91138da031c