Back socket
Summary:
Whew this is a doozy - This establishes the backchannel via websockets in the first commit, and then shows a working implementation in the form of the keyboard functionality in the second commit.

This includes a demo in the demo file of using these callbacks to set up a simple notepad-like pane. Seems to work:
![screen shot 2018-02-06 at 3 26 19 pm](https://user-images.githubusercontent.com/1276867/35883833-abd5e250-0b56-11e8-8843-3e91eaaa74cd.png)

Callback implementation format is described in the README, essentially allows registering one callback per pane, and then that function can case on the different event types if desired.

If desire appears, we'll also move to support handling events by event type rather than just by pane, but IDK if there's a strong enough case for that.
Closes https://github.com/facebookresearch/visdom/pull/260

Differential Revision: D6925327

Pulled By: JackUrb

fbshipit-source-id: 1c0bfb16763daee0a449971a462c16b1caec96c8