Video fix
Summary:
Turns out the video format coming in reads H/W in the form of tensors, but the documentation for setting those bounds specifies that they are to be set in W/H and then fed in H/W. What a bother... well it's fixed now.

Fixes https://github.com/facebookresearch/visdom/issues/209.
Closes https://github.com/facebookresearch/visdom/pull/267

Differential Revision: D6938129

Pulled By: JackUrb

fbshipit-source-id: a3a7e52d0b61e4f0ba16623d509d5aef461dc0c8