Better demo information.
Summary:
It isn't necessarily clear that our demo is not set up to work with the pip version as much as it's to work with the github version. For those who run into this issue while running our demo, this should now print an error when the demo fails.

As we test the demo before pushing, the demo failing is almost always going to be fixed by updating visdom.
Closes https://github.com/facebookresearch/visdom/pull/269

Differential Revision: D6981059

Pulled By: JackUrb

fbshipit-source-id: 19f30d0c5a7ec0cd959a4f27d223ffe047156d4e