Expose get api for getting data for existing windows.
Summary:
Creates a DataHandler to expose an endpoint that allows a user to get window data that exists on the server from the `visdom` class.

Adds method to the visdom class that calls the data handler.

Updates docs with the new function

Quickly cleans up the server file b/c it was throwing linting errors

Testing:
![screen shot 2018-02-21 at 11 30 03 am](https://user-images.githubusercontent.com/1276867/36492827-1203bf06-16fc-11e8-9dfb-ac476c81d0e4.png)

Closes #266 by implementing the desired feature.
Closes https://github.com/facebookresearch/visdom/pull/272

Differential Revision: D7055360

Pulled By: JackUrb

fbshipit-source-id: 739f230d017874dfdf40f6eef0bf5531364d4e9a