Better plotly feature parity
Summary:
Better supports general plotly features by adding opts to specify additional layout opts that we don't manage and additional trace options that we don't already provide.

On the downside it means there are now 2 ways to accomplish most of the things that we directly support, but those that use the plotly way are already versed enough in the plotly formatting to not necessarily need to use our implementation anyways, and our implementation is still better for people unfamiliar with plotly.

Closes #201, as error bars are implemented using trace options in plotly, so they are now possible for us as well.
Closes #197, as this is a layout option that we can now pass along.
Closes #200 for the same reason as above.

Prevents future "plotly does it but visdom doesn't" feature requests.
Closes https://github.com/facebookresearch/visdom/pull/275

Differential Revision: D7431423

Pulled By: JackUrb

fbshipit-source-id: e2f06aab366062b6ebcc54d6a3a50680ad683730