Image pan and zoom while allowing page to scroll
Summary:
I was getting annoyed when scrolling down would result in an image getting zoomed in. Also, sometimes, when I would scroll up, the image would zoom out to size zero and the only way to reset was to reload the page. Additionally, there was no way to pan the image after a zoom.

So I did the following:

* Shift + scroll = pan (shift + two finger move on trackpad)
* Ctrl + scroll = zoom (ctrl + two finger move OR pinch zoom on trackpad)
* Allow pan & zoom when on the Pane, not just on the image
* Double click on pane (not just image) to reset image
* Reset button in pane bar to reset image
Closes https://github.com/facebookresearch/visdom/pull/308

Differential Revision: D7604351

Pulled By: JackUrb

fbshipit-source-id: da6dcbf597ee4c7b5f8387f6d738cf7def8b0ff1