Making /env/* and /compare/* get-able again
Summary:
Somewhere along the way, the ability to have direct addressing to particular environments was broken. Comparing environments was introduced without direct addressing. This restores the ability to go directly to an env by url and introduces the same ability for compares as well.

Also fixes a webpack bug which was putting the main.js file into the wrong place.

Fixes #273 (finally)
Closes https://github.com/facebookresearch/visdom/pull/317

Differential Revision: D7674304

Pulled By: JackUrb

fbshipit-source-id: ba4d227e41d78a12a5b28ba3b4369030057a4078