Noisy cast for numerical titles (fixes #353)
Summary:
Added a noisy cast for the case of numerical titles ONLY, this means that things like `opts={'title'=torch.tensor(10)}` are not covered here and are catched instead by an assertion error. A direct cast for these cases would depend on the string representation of the tensor/ndarray, which I don't think is generally desired. However adding a special cast for the case of scalar values  might be worth it.
I also didn't know if/where to put tests other than the demo, which I don't think is a good idea since its better that users give strings as titles from the beginning.
Finally, I've never used lua before, a quick glance to the api showed me that the data is sent to the tornado server which unpacks the lua values into python, however I didn't quite get if it then calls the corresponding function in python or not.
Closes https://github.com/facebookresearch/visdom/pull/360

Differential Revision: D8102332

Pulled By: JackUrb

fbshipit-source-id: 521b97e670484f7d567b5e2c9d54a2bc79175503