Add pillow-simd as requirement if it is installed (#368)
Summary:
As discussed in #367, this PR allows visdom to work with Pillow-simd if the end user has it installed.
What this PR does, is that upon installation it checks if pillow-simd is installed and adds it to the requirements if it is. Otherwise, regular pillow is added.
Pillow-simd is a drop-in replacement that uses the same namespace, so no further changes are required.

__Disclaimer:__
I did not figure this one out on my own, but saw this solution in [pytorch/vision](https://github.com/pytorch/vision/pull/522).
This solution might have some problems I am not aware of, as I am no guru when working with setuptools. One thing I do know, is that this will cause a problem when creating pip wheels, so that should be taken into consideration. _(I dont think that this is a concern for this project)_
Closes https://github.com/facebookresearch/visdom/pull/368

Differential Revision: D8314365

Pulled By: JackUrb

fbshipit-source-id: 9cc52eee5ad129263136a7d6c6cfa86e288e89eb