Local logging to file, replay from logs (#375)
Summary:
Implements #321. Basically visdom would simply dispose of logs that failed, and if a user closed panes on the server there was no recourse to getting them back. This introduces a way to store the logs on the local visdom client's machine as well as the remote visdom server, ensuring the data remains safe. It also allows replaying of the data.

Basic implementation centers around logging the endpoint and json of every plotting event to a file, and then that file can later be replayed with that info. One gotcha is that all plots have to know what the window id was that they were plotted with in order to make sure that recorded updates to those plots will properly update the correct plot. This does mean that logs cannot be replayed multiple times in the same environment.

Demo:
![visdomreplaylog](https://user-images.githubusercontent.com/1276867/41258722-36fd1c36-6d9f-11e8-9f4a-81b566bad54c.gif)

This PR also includes documentation for some of the Visdom object params

Closes #321
Closes https://github.com/facebookresearch/visdom/pull/375

Differential Revision: D8480260

Pulled By: JackUrb

fbshipit-source-id: ae4a7f21660e1b77ae0823e759fde59961cdee41