Fixes modal default values (#374)
Summary:
By providing default values for the dropdowns in the modals, we prevent issues where the React component was in a degenerate state, which was causing issue #292. These values are pulled from the current state, so we are sure that the default values we provide will always be valid to render.

Fixes #292.
Closes https://github.com/facebookresearch/visdom/pull/374

Differential Revision: D8611391

Pulled By: JackUrb

fbshipit-source-id: 52b8e418e27c09375c1b31714eaa301228bf4a06