Fix 'plotlyplot' for numpy arrays. (#403)
Summary:
I realize that the `plotlyplot` function will fail if the Plotly figure's attributes are numpy arrays instead of plain lists, since the numpy arrays cannot be serialized to JSON in `_send`.

I thought the easiest solution is to make use of the Plotly Python client's `PlotlyJSONEncoder`, which properly deals with that as well as proper handling of NaNs and Infs. I load the JSON back as a Python dict so `_send` can add other attributes to it before JSON-serializing the entire message.

With this, `plotly.figure_factory.create_distplot` works:

<img width="494" alt="screenshot 2018-06-25 18 05 11" src="https://user-images.githubusercontent.com/987837/41878243-5376eb68-78a2-11e8-9d7c-2398918b34b3.png">
Closes https://github.com/facebookresearch/visdom/pull/403

Differential Revision: D8662070

Pulled By: JackUrb

fbshipit-source-id: 0489197bc2ddbd44a60380fb8695489c4494f1b7