Allows 'append' on non-existing plot for option re-use (#391)
Summary:
Closes #100.

People want a blankboard so that they can use the same command in every call of an evaluation loop to plot rather than needing to instantiate the graph before they can update. Existing workarounds are doing a regular plot call on the first iteration and appends for all others, or initializing the plot with a garbage value (like 0,0 on a line) and then appending from that point. Having an option that can create or append solves the issue.

This leverages `win_exists` to determine whether a particular call should be a create or append when passed the `'createappend'` option. It then does the correct action. Tested for both line and scatter, and existing behavior is maintained.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/391

Differential Revision: D8896449

Pulled By: JackUrb

fbshipit-source-id: 45ba7a2ecc0188deecd39df75a68b8dfe795f8fc