Scattergl (WebGL) support is added (#431)
Summary:
<!--- Provide a general summary of your changes in the Title above -->

<!--- Describe your changes in detail -->
Plotly supports `type='scattergl'` for scatter plots which makes use of WebGL. Ability to use this feature is added. A demo is added for creating and updating a WebGL based line plot. Although I also updated the documentation, it _could_ be better. As I can't write Lua I didn't make any changes on Lua scripts, but if the parameter `opts` is passed directly to Python code, then there is no change necessary.

<!--- Why is this change required? What problem does it solve? -->
When the number of points is very high, the visdom window freezes and/or crashes and/or is very slow. Using scattergl, plotly uses WebGL instead of SVG which works better for these kinds of plots.
<!--- If it fixes an open issue, please link to the issue here. -->
See issue https://github.com/facebookresearch/visdom/issues/427 that I created.

<!--- Please describe in detail how you tested your changes. -->
A line is added to demo.py which creates a plot and appends new data onto. It works without any problems.
<!--- Include details of your testing environment, experiments you ran to see how -->
<!--- your change affects existing areas of the code and their behaviors, etc. -->
One must add `opts.webgl=True` to create **and** update the plot.
<!--- One method of testing is to run the `demo.py` script from the examples -->
<!--- both on your branch and a clean branch and ensure that none of the functionality -->
<!--- appears different. Be sure to install from source when testing. -->

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [ ] For JavaScript changes, I have re-generated the minified JavaScript code.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/431

Differential Revision: D9086327

Pulled By: JackUrb

fbshipit-source-id: fdc2637341e34e39574cf81070ee4111d49be209