Added mouse location display (#459)
Summary:
Added a new feature on ImagePane : the x/y coordinates of the mouse w.r.t. the top-left corner of the image is displayed on the bottom-left corner of the image (only when the mouse is over the ImagePane)

I messed up a little with my local/upstream/origin repos so there are many commits attached to this PR, but there are only few new lines spread in 3 source files (Pane.j, ImagePane.js + css)

Image coordinates are useful when you work on object localization/detection in images.

Tested using demo.py

![screenshot from 2018-08-29 19 39 36](https://user-images.githubusercontent.com/7393801/44805003-59e5b980-abc3-11e8-90e5-b11a7d6c36c1.png)

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] New feature (non-breaking change which adds functionality)

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] For JavaScript changes, I have re-generated the minified JavaScript code.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/459

Differential Revision: D9582792

Pulled By: JackUrb

fbshipit-source-id: ec667e08140a172019333c87995afb657c5562de