Allows CompareEnv to be able to handle updates (#468)
Summary:
Before there was no casing for CompareEnv based functionality to be able to handle an updated pane, as the pane content is generated by the server and new requests would just pass through. This firstly allows socket subscriptions to check a set of envs rather than just a singular environment. Then from the JS side, if an updated panel is received when currently in the compare envs setup, instead of rendering the updated panel it requests a new set of compare panes from the server.

Fixes https://github.com/facebookresearch/visdom/issues/435

Created two environments with the same plot, ensured that in compare mode they would properly send a new compare request to the server every time a plot was updated that was in either of the compare environments. Ensured that no such request would be made for a plot in a non-selected env, and that no such request happens when only selecting a singular plot.

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] For JavaScript changes, I have re-generated the minified JavaScript code.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/468

Differential Revision: D9619860

Pulled By: JackUrb

fbshipit-source-id: 9ee4bed6f34dd1dc541d182958b039de37b6f9fe