Allow scatter plots with more legend items than number of classes (#498)
Summary:
Change two assert statements in the scatter function from == to >=. This is done in order to -
- allow more legend items than number of classes
- allow the label 1 to be missing and assert that the minimum label is not less than 1

Closes #480
Also allows scatter function calls with label 1 missing. This could happen in a case in which the label 1 is not included in the scatter function call

This has been tested by making 3d scatter plots by modifying demo.py. The 3d scatter demo in demo.py contains 2 legend items. I have increased it to 3 legend items and then made 3 plots -
- with the same labels as demo.py, ie, [1, 2]
- with labels [2, 3]
- with labels [1, 2]
Screenshot is attached below for all 3 plots.

In addition to this, I have tested the demo.py file on both my branch and a clean branch and everything is the same.

![image](https://user-images.githubusercontent.com/10346824/47455069-ea451280-d79e-11e8-8a76-505a53ae938b.png)

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] For JavaScript changes, I have re-generated the minified JavaScript code.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/498

Reviewed By: JackUrb

Differential Revision: D10853225

Pulled By: vtantia

fbshipit-source-id: 09f60112ab560f12ddfdc019f69960c678d7b7ec