Handle silently (warn) when trying to load corrupt env json. (#508)
Summary:
Fix issue #507.

Throw a warning when (instead of crashing) when failing to load an environment json.

Currently the visdom server fails to start when trying to load a corrupt json environment file. The thrown exception does not indicate the name of the corrupt file. This fix will log a warning stating the corrupt json file and continue with starting the server.

Tested on several corrupted json files.

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [X ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [X ] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] For JavaScript changes, I have re-generated the minified JavaScript code.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/508

Differential Revision: D13084953

Pulled By: JackUrb

fbshipit-source-id: d42dd235bcfd4ade521b72a1865860581fd56440