Bundling visdom dependencies in distribution package (#543)
Summary:
We are now allowed to bundle the frontend requirement scripts into the visdom distribution, so this makes it official. Also bumps the version up to 0.1.8.6 in order to release a new version.

This should prevent any problems that people have with downloading visdom dependencies. If you can acquire visdom, it should come with all the required scripts to run.

I created a distribution with `setup.py`, uninstalled visdom, and then reinstalled from the dist. When doing so I didn't need to download any additional packages.

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] For JavaScript changes, I have re-generated the minified JavaScript code.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/543

Differential Revision: D13643678

Pulled By: JackUrb

fbshipit-source-id: 64cf5709dbb8d589c1541d0823b4783256d0ee03