Remove non-ASCII characters in README.md
Why:

Reading a file with non-ascii characters will raise a UnicodeDecodeError
if the system locale is not set to UTF-8. Note that the characters in
question are not coherent with the rest of the documentation and should
be '`' rather than '‘'.

How:

Simply replace '‘' by '`', but also encode the `README.md` file while
reading in `setup.py` to avoid further problems if non-ascii were to be
added again.