Remove non-ASCII characters in README.md (#544)
Summary:
Simply replace characters ‘ by \' in `README.md` and encode the `README.md` file in `setup.py` to avoid further problems if non-ASCII characters were to be added again.

Reading a file with non-ascii characters will raise a UnicodeDecodeError if the system locale is not set to UTF-8. Note that the characters in question are not coherent with the rest of the documentation and should be ` rather than ‘.

Tested in a container where locale is not UTF-8.

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] For JavaScript changes, I have re-generated the minified JavaScript code.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/544

Differential Revision: D13658795

Pulled By: JackUrb

fbshipit-source-id: a04b658a7775db9acbab35f8888cbe0f4365631f