Fix clear button 1px-shifting glitch (#547)
Summary:
The clear button in the main toolbar in the visdom UI was experiencing a 1px layout jump when the tooltip was activated.

The reason for this was that the tooltip DOM element was being placed as a child in the bootstrap button group. The bootstrap button group has certain assumptions for its children, which was being violated by this DOM injection. (Specifically, the `.btn + .btn` CSS selector wasn't being applied to the last button in the button group since it wasn't immediately preceded by a sibling `.btn` element after the tooltip DOM injection.)

The fix, which is to inject the tooltip into the DOM at the base of the document instead (i.e. the `body` element), was already present in the code, but was being overridden accidentally. This PR fixes the unintentional override of the global tooltip config. It also re-introduces a delay animation that was previously set but also being overridden.

Without this fix, the button group with the clear button experiences a 1px layout jump when the tooltip is invoked, which is undesirable. Additionally, global tooltip configuration options do not propagate as expected.

I compared the behavior of the previous code with the fixed code, and confirmed that the behavior is the same for each click of the UI element. The only caveat is that the fix also reintroduces a slight delay animation which was previously

![bugfix](https://user-images.githubusercontent.com/425059/51149736-48bae380-1818-11e9-8892-5db05d4974c5.png)

- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

- [X] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [X] For JavaScript changes, I have re-generated the minified JavaScript code.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/547

Reviewed By: JackUrb

Differential Revision: D13674058

Pulled By: pringshia

fbshipit-source-id: 786e0feabd470c8ec66d50acef37e0bee773316a