Fix issues with audio/video objects on Windows (#540)
Summary:
Fix issues related to assumptions about where temporary files can be stored, which caused several types of failures on Windows for audio and video objects.

This allows the demo script to exercise the full range of capabilities on Windows.  Furthermore, the demo script would previously fail in the audio example, which caused the script to terminate prematurely and caused some initial confusion as to why callbacks were not working as expected.

Ran demo script and observed that it now produces audio and video objects on Windows, both from tensor data and from files.

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] For JavaScript changes, I have re-generated the minified JavaScript code.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/540

Differential Revision: D14601968

Pulled By: JackUrb

fbshipit-source-id: edfb2f1bf26022625f7ce112eee7aeaefd880234