Logging tracebacks during errors on the server (#585)
Summary:
Right now when an error occurs on the visdom server we print the error, but no details about what actually happened or the traceback. This makes it particularly hard to debug when something goes wrong and where the error happened when it happens. This change adds a print for the exception information so that we can see what happened.

As in the case of #560, server issues are currently very hard to debug without this information.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/585

Differential Revision: D14605982

Pulled By: JackUrb

fbshipit-source-id: b42796c9bb69dca0a13e3771f2258995242ebdff