Warning on None opts that can't be None (#586)
Summary:
Certain options, namely `title`, cause a crash when a None value is provided to the opts dict. This PR makes it so these opts are removed from the dict, allowing the plot to not crash everything. I warn when this occurs though, as it is more likely to be the result of a bug than someone intentionally passing None to that opt value.

Fixes #563.

```
>>> vis.line(X=np.array((2,)), Y=np.array(((2,))), opts={'title': None})
WARNING:visdom:None-incompatible opt title was provided None value and was thus ignored
'window_37317b92756306'
```
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/586

Differential Revision: D14648057

Pulled By: JackUrb

fbshipit-source-id: 88d0c5d6fc20fab26fe18b6a1db2f3af5c79ab19