Moving layout and state management to server app init
Summary:
It hasn't really made sense that a socket handler was handling initialization of some of the application state, so now that portion is initialized in the app startup. This was a problem in the past, as it meant that layouts weren't loaded until the first time a socked opened, which doesn't really make sense.

This has the benefit of allowing something that is starting up an application to be able to override these values before deploying.

Reviewed By: pringshia

Differential Revision: D15481131

fbshipit-source-id: 0194be83c05bf856f9dc19ad18a4f590dc281e66