Fixing bin_packer filetype
Summary: This file has been mislabeled forever. On unsecured http setups, the browser has been happy to digest it in whatever form it came, but when trying to adapt to running over https via an SSL wrapper, the frontend fails to load this file. This is caused by the fact that tornado serves it up using a different MIME type than what's expected due to the improper filename. This patch fixes the name.

Reviewed By: pringshia, lvdmaaten

Differential Revision: D15542737

fbshipit-source-id: 0871913f62e393c3b5ab89c7e422cac3d7c2bcf1