Append events in local log fix (#626)
Summary:
As surfaced in #624, there are some issues in the interactions between local logging and append, especially when working in offline mode. This is the fix.

Fixes #624

## Description
The core of the interaction here that wasn't working properly was that append events were not being properly handled due to `win_exists` not successfully returning that a window had been created. This now stores append events into the logs as such both when offline and when `win_exists` returns none (failed connection). In order to handle the fact that it's possible that append events may be sent to the server on a replay_log event, I've also cased `UpdateHandler` to be able to create a window during an append event if none existed already.

## How Has This Been Tested?
Ran the test case in #624

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist:

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] For JavaScript changes, I have re-generated the minified JavaScript code.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/626

Differential Revision: D15603058

Pulled By: JackUrb

fbshipit-source-id: 40b26335326cdb79046afec4e4c00291b426e867