Adding optional polling-based comms to frontend (#633)
Summary:
## Description
Creates piping to use the same API for both socket-based and request-based communication between the frontend client and the server.

## Motivation and Context
Up until now, the only way to communicate with a visdom backend has been through sockets. Unfortunately, in some network configurations, socket-based communications are not scalable or not feasible. To this end, this PR introduces a method to have both.

## How Has This Been Tested?
Ran the demo both with `-use_frontend_client_polling` and without it.

## Screenshots (if appropriate):
With:
![Screen Shot 2019-06-12 at 6 52 07 PM](https://user-images.githubusercontent.com/1276867/59391980-03a5c980-8d44-11e9-8ac9-20d5d5e516f6.png)
Without:
![Screen Shot 2019-06-12 at 6 50 23 PM](https://user-images.githubusercontent.com/1276867/59391982-056f8d00-8d44-11e9-8db0-9981228582dc.png)

Note that the batching element of the query-based version has led to a different ordering of windows. In observation this has to do with the number of times repack is called on different batch sizes.

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist:

- [x My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] For JavaScript changes, I have re-generated the minified JavaScript code.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/633

Differential Revision: D15806326

Pulled By: JackUrb

fbshipit-source-id: 067f321be7946bc01953d65f4c6e2ecfa8cc2297