Adds support for transferring env state to/from client (#631)
Summary:
## Description
Currently Visdom handles serialization with `save`, but this means that log files end up in a different place to the rest of the project. This PR adds a `set_window_data` function that, together with the existing `get_window_data`, can be used to serialize data on the client.

It also fixes `get_window_data`/`close` by changing `win=None` to `win='all'`, as `_send` assumes `None` means 'create a new window' and helpfully fills in `win=None` with a random string.

## How Has This Been Tested?
I've manually checked it works for setting the data of environments that don't currently exist on the server. It doesn't yet work for updating environments that already exist - what's the right way to trigger a reload on all the clients?

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist:

- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [ ] For JavaScript changes, I have re-generated the minified JavaScript code.
Pull Request resolved: https://github.com/facebookresearch/visdom/pull/631

Differential Revision: D15905655

Pulled By: JackUrb

fbshipit-source-id: d2347bdeadf97b0836a0ed95cad09886c84eeac6